Add config volume example to template
This example shows how one would get a custom config file into a
specific container.
This method uses an external volume independent of the Instant OpenHIE
volume that will hold the custom config.
This volume is created before the component container is created as the
config in this scenario needs to be present before the container starts
up.

DIO-28